refactor: summarize Node types to block/node/type.ts
