+++ src/block/node/CommandLineNode.ts
@@ -1,14 +1,9 @@
 import { createNodeParser } from './creator'
 
+import type { CommandLineNode } from './type'
 import type { NodeCreator } from './creator'
 
 const commandLineRegExp = /^[$%] .+$/
-
-export interface CommandLineNode {
-  type: 'commandLine'
-  symbol: string
-  text: string
-}
 
 const createCommandLineNode: NodeCreator<CommandLineNode> = (target: string) => {
   const symbol = target[0]
