+++ src/block/node/HelpfeelNode.ts
@@ -1,13 +1,9 @@
 import { createNodeParser } from './creator'
 
+import type { HelpfeelNode } from './type'
 import type { NodeCreator } from './creator'
 
 const helpfeelRegExp = /^\? .+$/
-
-export interface HelpfeelNode {
-  type: 'helpfeel'
-  text: string
-}
 
 const createHelpfeelNode: NodeCreator<HelpfeelNode> = target => ({
   type: 'helpfeel',
