+++ src/block/node/InternalLinkNode.ts
@@ -1,15 +1,9 @@
 import { createNodeParser } from './creator'
 
+import type { InternalLinkNode } from './type'
 import type { NodeCreator } from './creator'
 
 const internalLinkRegExp = /\[\/?[^[\]]+\]/
-
-export interface InternalLinkNode {
-  type: 'link'
-  pathType: 'root' | 'relative'
-  href: string
-  content: string
-}
 
 const createInternalLinkNode: NodeCreator<InternalLinkNode> = target => {
   const href = target.substring(1, target.length - 1)
