+++ src/block/node/StrongIconNode.ts
@@ -1,15 +1,10 @@
 import { createNodeParser } from './creator'
 import { generateIconNodeCreator } from './IconNode'
 
+import type { StrongIconNode } from './type'
 import type { NodeCreator } from './creator'
 
 const strongIconRegExp = /\[\[[^[\]]*\.icon(?:\*\d+)?\]\]/
-
-export interface StrongIconNode {
-  type: 'strongIcon'
-  pathType: 'root' | 'relative'
-  path: string
-}
 
 const createStrongIconNode = generateIconNodeCreator('strongIcon') as NodeCreator<StrongIconNode>
